+++ sky/skylet/providers/kubernetes/__init__.py
@@ -1,2 +1,1 @@
-from sky.skylet.providers.kubernetes.utils import get_head_ssh_port, get_port
 from sky.skylet.providers.kubernetes.node_provider import KubernetesNodeProvider
