+++ sky/skylet/providers/kubernetes/node_provider.py
@@ -7,8 +7,7 @@ from uuid import uuid4
 
 from sky.adaptors import kubernetes
 from sky.skylet.providers.kubernetes import config
-from sky.skylet.providers.kubernetes import get_head_ssh_port
-from sky.skylet.providers.kubernetes import utils
+from sky.utils import kubernetes_utils
 from ray.autoscaler._private.command_runner import SSHCommandRunner
 from ray.autoscaler.node_provider import NodeProvider
 from ray.autoscaler.tags import NODE_KIND_HEAD, TAG_RAY_CLUSTER_NAME, TAG_RAY_NODE_KIND
@@ -50,7 +49,7 @@ class KubernetesNodeProvider(NodeProvider):
         self.cluster_name = cluster_name
 
         # Kubernetes namespace to user
-        self.namespace = utils.get_current_kube_config_context_namespace()
+        self.namespace = kubernetes_utils.get_current_kube_config_context_namespace()
 
         # Timeout for resource provisioning. If it takes longer than this
         # timeout, the resource provisioning will be considered failed.
@@ -116,7 +115,7 @@ class KubernetesNodeProvider(NodeProvider):
         # TODO(romilb): Implement caching here for performance.
         # TODO(romilb): Multi-node would need more handling here.
         cluster_name = node_id.split('-ray-head')[0]
-        return get_head_ssh_port(cluster_name, self.namespace)
+        return kubernetes_utils.get_head_ssh_port(cluster_name, self.namespace)
 
     def internal_ip(self, node_id):
         pod = kubernetes.core_api().read_namespaced_pod(node_id, self.namespace)
