+++ sky/templates/kubernetes-ray.yml.j2
@@ -189,7 +189,6 @@ available_node_types:
               memory: {{memory}}G
               nvidia.com/gpu: {{accelerator_count}}
             limits:
-              cpu: {{cpus}} # TODO(romilb): Ideally, we should not specify CPU limits here to allow jobs to use idle CPU cycles. However, CPUs limits are required for Ray CPU count initialization.
               nvidia.com/gpu: {{accelerator_count}} # Limits need to be defined for GPU requests
   ray_worker_default:
     # Minimum number of Ray workers of this Pod type.
@@ -252,7 +251,6 @@ available_node_types:
               memory: {{memory}}G
               nvidia.com/gpu: {{accelerator_count}}
             limits:
-              cpu: {{cpus}} # TODO(romilb): Ideally, we should not specify CPU limits here to allow jobs to use idle CPU cycles. However, CPUs limits are required for Ray CPU count initialization.
               nvidia.com/gpu: {{accelerator_count}} # Limits need to be defined for GPU requests
 
 setup_commands:
