+++ sky/utils/__init__.py
@@ -1,2 +1,1 @@
 """Utility functions."""
-from sky.skylet.providers.kubernetes import utils as kubernetes_utils
